Remove deprecated methods; Add additional extension methods for Where clause
