+++ version.props
@@ -1,0 +1,9 @@
+<Project>
+  <PropertyGroup>
+    <VersionPrefix>3.2.0</VersionPrefix>
+    <VersionSuffix>preview</VersionSuffix>
+    <PackageVersion Condition="'$(VersionSuffix)' == '' ">$(VersionPrefix)</PackageVersion>
+    <BuildNumber Condition="'$(BuildNumber)' == ''">t0000</BuildNumber>
+    <VersionSuffix Condition="'$(VersionSuffix)' != '' AND '$(BuildNumber)' != ''">$(VersionSuffix)-$(BuildNumber)</VersionSuffix>
+  </PropertyGroup>
+</Project>
